폴더 구조 수정 및 idea  제외, clsx 추가, max-len:180으로 변경 및 적용
