+++ .gitignore
@@ -13,6 +13,9 @@
 # testing
 /coverage
 
+# intellij
+.idea/
+
 # next.js
 /.next/
 /out/
@@ -39,5 +42,3 @@ yarn-error.log*
 # typescript
 *.tsbuildinfo
 next-env.d.ts
-/app/writeform/page.tsx
-/app/writeform/tiptap-editor.css
