+++ app/globals.css
@@ -24,3 +24,52 @@ body {
   color: var(--foreground);
   font-family: Arial, Helvetica, sans-serif;
 }
+
+/* swiper-custom.css 등 원하는 위치에 추가 */
+.swiper-pagination {
+  position: absolute;
+  left: 0;
+  right: 0;
+  bottom: 16px; /* 이 값을 통해 원하는 만큼 아래로! */
+  display: flex;
+  justify-content: center;
+  z-index: 10;
+}
+.swiper {
+  padding-bottom: 40px !important; /* 페이지네이션이 겹치지 않도록 바닥에 여유 주기 */
+}
+/* swiper-custom.css */
+.swiper-pagination-bullet {
+  background: #cbd5e1; /* Tailwind gray-300 */
+  opacity: 1;
+}
+
+.swiper-pagination-bullet-active {
+  background: #3b82f6; /* Tㅈailwind blue-500 or 원하는 색 */
+}
+
+/* swiper-custom.css */
+.swiper .swiper-pagination,
+.swiper .swiper-pagination-bullets,
+.swiper .swiper-pagination-bullets.swiper-pagination-horizontal,
+.swiper .swiper-pagination-bullet {
+  display: none !important;
+}
+
+/* swiper-custom.css - 이미 존재하는 컴포넌트에 맞춰 추가/수정 */
+.mt-4.flex.justify-center.swiper-pagination-bullet {
+  /* 점이 나타나는 영역 스타일 */
+  gap: 15px;
+}
+.mt-4.flex.justify-center .swiper-pagination-bullet {
+  width: 10px;
+  height: 10px;
+  background: #ccc;
+  border-radius: 50%;
+  opacity: 1;
+  margin: 0 4px;
+  transition-duration: 0.3s;
+}
+.mt-4.flex.justify-center .swiper-pagination-bullet-active {
+  background: #3b82f6; /* 활성화된 점 별도 강조 */
+}
