+++ postcss.config.mjs
@@ -1,3 +1,4 @@
+/** @type {import('tailwindcss').Config} */
 const config = {
   plugins: ['@tailwindcss/postcss'],
 };
